move plot_one_box to utils.py
