+++ efficientdet_test.py
@@ -13,7 +13,7 @@ import cv2
 import numpy as np
 
 from efficientdet.utils import BBoxTransform, ClipBoxes
-from utils.utils import preprocess, invert_affine, postprocess, STANDARD_COLORS, standard_to_bgr, get_index_label
+from utils.utils import preprocess, invert_affine, postprocess, STANDARD_COLORS, standard_to_bgr, get_index_label, plot_one_box
 
 compound_coef = 0
 force_input_size = None  # set None to use default size
@@ -78,18 +78,6 @@ with torch.no_grad():
                       regressBoxes, clipBoxes,
                       threshold, iou_threshold)
 
-def plot_one_box(img, coord, label=None, score=None, color=None, line_thickness=None):
-    tl = line_thickness or int(round(0.001 * max(img.shape[0:2])))  # line thickness
-    color = color
-    c1, c2 = (int(coord[0]), int(coord[1])), (int(coord[2]), int(coord[3]))
-    cv2.rectangle(img, c1, c2, color, thickness=tl)
-    if label:
-        tf = max(tl - 2, 1)  # font thickness
-        s_size = cv2.getTextSize(str('{:.0%}'.format(score)),0, fontScale=float(tl) / 3, thickness=tf)[0]
-        t_size = cv2.getTextSize(label, 0, fontScale=float(tl) / 3, thickness=tf)[0]
-        c2 = c1[0] + t_size[0]+s_size[0]+15, c1[1] - t_size[1] -3
-        cv2.rectangle(img, c1, c2 , color, -1)  # filled
-        cv2.putText(img, '{}: {:.0%}'.format(label, score), (c1[0],c1[1] - 2), 0, float(tl) / 3, [0, 0, 0], thickness=tf, lineType=cv2.FONT_HERSHEY_SIMPLEX)
 
 
 def display(preds, imgs, imshow=True, imwrite=False):
